core: Pass return code to arch_cpu_restore
Some architectures, so far ARM, may prefer to jump directly to the
target Linux context from arch_cpu_restore. In this case we need to have
the return code at hand as well. Extend the parameter list accordingly
and document the possibility that arch_cpu_restore does not return to
the caller.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>